refactor: `Place` 클래스 이름 변경 및 파일 이동
- `Place` 클래스의 이름을 `PlaceUiModel`로 변경
- `PlaceUiModel` 및 `PlaceCategory`를 `uimodel` 패키지에서 `model` 패키지로 이동
- `PlaceUiModel`에 `isBookmarked` 프로퍼티 추가
