+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/PlaceListFragment.kt
@@ -18,7 +18,7 @@ class PlaceListFragment :
         super.onViewCreated(view, savedInstanceState)
         binding.rvPlaces.adapter =
             PlaceListAdapter().apply {
-                submitList(DummyPlace.placeList)
+                submitList(DummyPlace.placeUiModelList)
             }
     }
 }
