+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/dummy/DummyPlace.kt
@@ -1,14 +1,14 @@
 package com.daedan.festabook.presentation.placeList.dummy
 
-import com.daedan.festabook.presentation.placeList.uimodel.Place
-import com.daedan.festabook.presentation.placeList.uimodel.PlaceCategory
+import com.daedan.festabook.presentation.placeList.model.PlaceCategory
+import com.daedan.festabook.presentation.placeList.model.PlaceUiModel
 
 object DummyPlace {
-    val placeList =
+    val placeUiModelList =
         buildList {
             repeat(100) { i ->
                 add(
-                    Place(
+                    PlaceUiModel(
                         id = i.toLong(),
                         imageUrl = "https://cdn-dantats.stunning.kr/prod/portfolios/2f38e380-1ddd-4916-bd91-2ea3e901ff50/contents/YuoATBxT6sxa7ey2.%EC%82%B0%EB%94%94-10110-%EB%B0%B0%ED%98%9C%EC%97%B0.jpg.small?q=80&f=webp&t=crop&s=3508x4961",
                         category = PlaceCategory.BOOTH,
