+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/model/PlaceCategory.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeList.uimodel
+package com.daedan.festabook.presentation.placeList.model
 
 enum class PlaceCategory {
     FOOD_TRUCK,
